--- lib/proforma/renderers.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-#
-# Copyright (c) 2019-present, Blue Marble Payroll, LLC
-#
-# This source code is licensed under the MIT license found in the
-# LICENSE file in the root directory of this source tree.
-#
-
-require_relative 'renderers/plain_text_renderer'
